refactor: modularize transactions feature
