+++ modules/transactions/components/NewTransaction.tsx
@@ -9,8 +9,8 @@ import Image from 'next/image'
 import { imageHelper } from '@/utils/image-helper'
 import { formatCurrency } from '@/utils/currency'
 import { fetcher } from '@/utils/fetcher'
-import { useTransactions } from '@/utils/useTransactions'
-import { useTransactionKinds } from '@/utils/useTransactionKinds'
+import { useTransactionKinds } from '@/modules/transactions/hooks/useTransactionKinds'
+import { useTransactions } from '@/modules/transactions/hooks/useTransactions'
 import { Loader } from '@/components/Loader'
 import Card from '@/public/static/images/card.png'
 import PixelsLight from '@/public/static/images/pixels-light.png'
