+++ modules/transactions/components/index.ts
@@ -1,0 +1,3 @@
+export { NewTransaction } from './NewTransaction'
+export { Statement } from './Statement'
+export { TransactionsPie } from './TransactionsPie'
