+++ modules/transactions/hooks/useTransactionKinds.ts
@@ -1,7 +1,7 @@
 import useSWR from 'swr'
 import { fetcher } from '@/utils/fetcher'
 import type { Kind } from '@/app/api/transaction-kinds/types'
-import { SWR_KEYS } from './swr-keys'
+import { SWR_KEYS } from '@/utils/swr-keys'
 
 export function useTransactionKinds() {
   const { data, error, isLoading } = useSWR<{ kinds: Kind[] }>(
